skipping mygene tests if they are stalling for too long
