+++ src/tests/test_ontologies_mygene.py
@@ -1,7 +1,9 @@
+from .conftest import skip_on_timeout
 from napistu.ontologies.mygene import create_python_mapping_tables
 from napistu.ontologies.constants import INTERCONVERTIBLE_GENIC_ONTOLOGIES
 
 
+@skip_on_timeout(5)
 def test_create_python_mapping_tables_yeast():
     """Test create_python_mapping_tables with yeast species."""
     # Test with a subset of mappings to keep test runtime reasonable
